% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.weeknumber.R
\name{as.weeknumber}
\alias{as.weeknumber}
\title{Coerce to week number class}
\usage{
as.weeknumber(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Coerce object to week number class.
}
\examples{
as.weeknumber(c(-1:1, 51:52, NA))
as.weeknumber("2000-W01")
as.weeknumber(as.Date("2000-12-28"))

}
