% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{internal}
\alias{internal}
\alias{scale_type.weeknumber}
\alias{internal}
\alias{type_sum.weeknumber}
\title{Internal}
\usage{
scale_type.weeknumber(x)

type_sum.weeknumber(x)
}
\description{
Internal
}
\keyword{internal}
