% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeknumber-package.R
\name{is_weeknumber}
\alias{is_weeknumber}
\title{Test for weeknumber vectors}
\usage{
is_weeknumber(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{weeknumber} vector, otherwise \code{FALSE}.
}
\description{
Tests whether an object inherits from the \code{weeknumber} class.
}
\examples{
x <- weeknumber(10)
is_weeknumber(x)
is_weeknumber(10)
}
