% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{seq.weeknumber}
\alias{seq.weeknumber}
\title{Generate Sequences of Week Numbers}
\usage{
\method{seq}{weeknumber}(from, to, by, length.out = NULL, along.with = NULL, ...)
}
\arguments{
\item{from}{A length-1 \code{weeknumber} vector giving the start of the
sequence.}

\item{to}{A length-1 \code{weeknumber} vector giving the end of the sequence.}

\item{by}{A length-1 numeric week increment.}

\item{length.out}{Desired length of the output sequence.}

\item{along.with}{Take the length from this object.}

\item{...}{Unused.}
}
\value{
A \code{weeknumber} vector.
}
\description{
Generate regular sequences of \code{weeknumber} values.
}
\details{
This method mirrors classed \code{seq()} methods such as \code{\link[=seq.Date]{seq.Date()}}. Supply
\code{from} and \code{to}, or one endpoint together with \code{length.out}/\code{along.with}.
When \code{by} is supplied, it should be a single numeric week increment.
}
\examples{
seq(make_weeknumber(2000, 1), make_weeknumber(2000, 9), by = 2)
seq(from = make_weeknumber(2000, 1), length.out = 3, by = 1)

}
