% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{as_weeknumber}
\alias{as_weeknumber}
\title{Coerce to week number class}
\usage{
as_weeknumber(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A \code{weeknumber} vector.
}
\description{
Coerce an object to a \code{weeknumber} vector.
}
\details{
Convert numbers, ISO week strings like \code{"2000-W01"}, factors, and date-time
objects to \code{weeknumber} values.
}
\examples{
as_weeknumber(c(-1:1, 51:52, NA))
as_weeknumber("2000-W01")
as_weeknumber(as.Date("2000-12-28"))
}
