% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arith.R, R/cast.R, R/internal.R, R/ptype.R
\name{internal}
\alias{internal}
\alias{vec_arith.weeknumber}
\alias{vec_cast.weeknumber}
\alias{vec_ptype2.weeknumber}
\title{Internal}
\usage{
\method{vec_arith}{weeknumber}(op, x, y, ...)

\method{vec_cast}{weeknumber}(x, to, ...)

\method{vec_ptype2}{weeknumber}(x, y, ...)
}
\description{
Internal functions.
}
\keyword{internal}
