% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_weeknumber}
\alias{scale_weeknumber}
\alias{scale_x_weeknumber}
\alias{scale_weeknumber}
\alias{scale_y_weeknumber}
\title{Scales for week numbers}
\usage{
scale_x_weeknumber(name = ggplot2::waiver(), breaks = ggplot2::waiver(),
  minor_breaks = ggplot2::waiver(), labels = ggplot2::waiver(),
  limits = NULL, expand = ggplot2::waiver(), oob = scales::censor,
  na.value = NA_real_, position = "bottom")

scale_y_weeknumber(name = ggplot2::waiver(), breaks = ggplot2::waiver(),
  minor_breaks = ggplot2::waiver(), labels = ggplot2::waiver(),
  limits = NULL, expand = ggplot2::waiver(), oob = scales::censor,
  na.value = NA_real_, position = "left")
}
\arguments{
\item{name}{See link for details.}

\item{breaks}{See link for details.}

\item{minor_breaks}{See link for details.}

\item{labels}{See link for details.}

\item{limits}{See link for details.}

\item{expand}{See link for details.}

\item{oob}{See link for details.}

\item{na.value}{See link for details.}

\item{position}{See link for details.}
}
\description{
See \code{\link[ggplot2]{scale_continuous}} for details.
}
