% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_weeknumber}
\alias{scale_weeknumber}
\alias{scale_x_weeknumber}
\alias{scale_y_weeknumber}
\title{Scales for week numbers}
\usage{
scale_x_weeknumber(
  name = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  minor_breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  expand = ggplot2::waiver(),
  oob = scales::censor,
  na.value = NA_real_,
  position = "bottom"
)

scale_y_weeknumber(
  name = ggplot2::waiver(),
  breaks = ggplot2::waiver(),
  minor_breaks = ggplot2::waiver(),
  labels = ggplot2::waiver(),
  limits = NULL,
  expand = ggplot2::waiver(),
  oob = scales::censor,
  na.value = NA_real_,
  position = "left"
)
}
\arguments{
\item{name, breaks, minor_breaks, labels, limits, expand, oob, na.value, position}{See link for details.}
}
\description{
See \code{\link[ggplot2:scale_x_continuous]{ggplot2::scale_x_continuous()}} for details.
}
