% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeknumber-package.R
\docType{package}
\name{weeknumber-package}
\alias{weeknumber-package}
\title{weeknumber: A Class for Week Numbers}
\description{
A class for week numbers, according to ISO 8601.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hmalmedal.github.io/weeknumber/}
  \item \url{https://github.com/hmalmedal/weeknumber}
  \item Report bugs at \url{https://github.com/hmalmedal/weeknumber/issues}
}

}
\author{
\strong{Maintainer}: Håkon Malmedal \email{hmalmedal@gmail.com}

}
\keyword{package}
