% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeknumber-package.R
\name{weeknumber}
\alias{weeknumber}
\alias{new_weeknumber}
\alias{is_weeknumber}
\title{Week number class}
\usage{
new_weeknumber(x = double())

weeknumber(x = double())

is_weeknumber(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Constructor, helper and test functions for the week number class.
}
