% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_week.R
\name{year_week}
\alias{year_week}
\title{Get year and week number}
\usage{
year_week(x)
}
\arguments{
\item{x}{An object}
}
\value{
A named list with two elements: \code{year} and \code{week}.
}
\description{
Get year and week number
}
\examples{
x <- as.weeknumber(c(-1:1, 51:52, NA))
year_week(x)

}
